refactor: 파일 패키지 구조 수정
