+++ backend/src/main/java/woorifisa/goodfriends/backend/phoneAuth/application/SmsService.java
@@ -11,6 +11,7 @@ import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
+import woorifisa.goodfriends.backend.phoneAuth.domain.CreateSignature;
 import woorifisa.goodfriends.backend.phoneAuth.dto.MessagesDto;
 import woorifisa.goodfriends.backend.phoneAuth.dto.request.SmsRequest;
 import woorifisa.goodfriends.backend.phoneAuth.dto.response.SmsResponse;
