+++ backend/src/main/java/woorifisa/goodfriends/backend/phoneAuth/domain/CreateSignature.java
@@ -1,4 +1,4 @@
-package woorifisa.goodfriends.backend.phoneAuth.application;
+package woorifisa.goodfriends.backend.phoneAuth.domain;
 
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
